// src/app/features/profiles/models/profiles.model.ts
export interface Profile {
  key:any,
  address: string;
  birthdate: Date;
  email: string;
  email2: string;
  first_name: string;
  gender: string;
  last_name: string;
  fullName: string;
  localid: number;
  loyalty_member_id: string;
  modified: Date;
  phone: string;
  photo: string;
  prefix: string;
  suffix: string;
}